fix: extract CLI resolution into agentv_cli.py wrapper, remove AGENT_EVAL_MODE from .env
AGENT_EVAL_MODE was a workflow preference stored alongside API keys in
.env — wrong place for non-secret config. Eval mode is now determined
by user instruction, defaulting to agent mode.

Also deduplicated _find_agentv() / _find_env_key() from three scripts
into a shared agentv_cli.py module that handles CLI resolution
deterministically.
