+++ .env.example
@@ -1,8 +1,5 @@
 # Copy this file to .env and fill in your credentials
 
-# Eval run mode (used by agentv-bench skill)
-AGENT_EVAL_MODE=agent # agent | cli
-
 # Azure OpenAI Configuration
 AZURE_OPENAI_ENDPOINT=https://your-endpoint.openai.azure.com/
 AZURE_OPENAI_API_KEY=your-openai-api-key-here
